+++ codec/asciinema/new.go
@@ -1,6 +1,7 @@
 package asciinema
 
 import (
+	"github.com/containerssh/geoip"
 	"github.com/containerssh/log"
 
 	"github.com/containerssh/auditlog/codec"
@@ -8,8 +9,9 @@ import (
 
 // NewEncoder Creates an encoder that writes in the Asciicast v2 format
 // (see https://github.com/asciinema/asciinema/blob/develop/doc/asciicast-v2.md)
-func NewEncoder(logger log.Logger) codec.Encoder {
+func NewEncoder(logger log.Logger, geoIPProvider geoip.LookupProvider) codec.Encoder {
 	return &encoder{
-		logger: logger,
+		logger:        logger,
+		geoIPProvider: geoIPProvider,
 	}
 }